take fee into account for spreadthresholdselector

+++ src/main/kotlin/at/gpro/arbitrader/entity/Exchange.kt
@@ -1,8 +1,6 @@
 package at.gpro.arbitrader.entity
 
-import java.math.BigDecimal
-
 interface Exchange {
     fun getName(): String
-    fun getFee(): BigDecimal
+    fun getFee(): Double
 }
