+++ src/main/kotlin/at/gpro/arbitrader/xchange/utils/ExchangeConverter.kt
@@ -1,13 +1,12 @@
 package at.gpro.arbitrader.xchange.utils
 
 import at.gpro.arbitrader.entity.Exchange
-import java.math.BigDecimal
 
 class ExchangeConverter {
-    fun convert(xchange: XchangeExchange, fee: BigDecimal) =
+    fun convert(xchange: XchangeExchange, fee: Double) =
         object : Exchange {
             override fun getName(): String = xchange.exchangeSpecification.exchangeName
-            override fun getFee(): BigDecimal = fee
+            override fun getFee(): Double = fee
             override fun toString(): String = getName()
         }
 }
